refactor and add build tags to eventually generate templates
