+++ cmd/resources/debugging.tmpl
@@ -1,0 +1,1 @@
+{{ printf "%#v" . }}
